+++ Changelog.md
@@ -9,6 +9,8 @@
 - Fix VisualNav routes not always redrawing properly
 - Fix DungeonMaps Zoom now gets properly saved between sessions
 - Fix DPI-scaling issue when /ub playsound is run
+- Fix VideoPatchFocus on login, no longer needs to be fiddled with to work after relogging
+- Fix Auto disable nametags while videopatch is enabled.
 
 ## 0.1.0 (2019-11-30) [UtilityBeltInstaller-0.1.0.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/d92c9e54fb61ab148bc2e1d2342ee204/UtilityBeltInstaller-0.1.0.exe)
 - Added AutoTinker and AutoImbue (thanks Schneebly)